git-svn-id: https://plantuml.svn.sourceforge.net/svnroot/plantuml/trunk@16 28b6c6c1-be0e-40f3-8fed-77912e9d39c1

+++ src/net/sourceforge/plantuml/skin/bluemodern/ComponentBlueModernParticipant.java
@@ -66,7 +66,7 @@ public class ComponentBlueModernParticipant extends AbstractTextualComponent {
 		shape.draw(g2d);
 		g2d.translate(-outMargin, 0);
 
-		getTextBlock().draw(g2d, (int) (outMargin + getMarginX1()), getMarginY());
+		getTextBlock().draw(g2d, outMargin + getMarginX1(), getMarginY());
 	}
 
 	@Override
